♻️ refactor(ICM-20948): use Port B / SSI 2

+++ Library/SSI/SSI0.h
@@ -4,9 +4,9 @@
 
 void SSI0_Init(uint32_t SYS_CLK, uint32_t SSI_CLK, SSI_MODE frameConfig, DATA_SIZE dataSize);
 
-void SSI0_Read(uint16_t initData, uint16_t *result, uint8_t length);
+void SSI0_Read(uint16_t initData, uint16_t *result, uint32_t length);
 
-void SSI0_Write(uint16_t *data, uint8_t length);
+void SSI0_Write(uint16_t *data, uint32_t length);
 
 void SSI0_StartTransmission(void);
